+++ main.js
@@ -33,7 +33,7 @@ function createWindow () {
 		slashes: true
 	}))
 	// Open the DevTools.
-	// mainWindow.webContents.openDevTools()
+	mainWindow.webContents.openDevTools()
 	// Emitted when the window is closed.
 	mainWindow.on('closed', function () {
 		// Dereference the window object, usually you would store windows